generic-organization/requirements.txt
Django==3.1.6
psycopg2-binary==2.8.5
djangorestframework
requests==2.25.1
#amqp added for bug introduced https://github.com/celery/py-amqp/issues/340
amqp==2.6.1
celery==4.3.0
marshmallow==2.20.0
zeep==3.4.0
redis==3.3.11
Pillow==6.0.0
mrz==0.5.8
django-pgcrypto==2.0.0
antidote==0.7.3
--index-url=http://pypi.etuitus.it --trusted-host pypi.etuitus.it
dizme-agent-pyclient==0.0.2
infocert-pyclients==0.0.1
